+++ lib/dp_exchange/webull.ex
@@ -108,9 +108,6 @@ defmodule DpExchange.Webull do
     {:get_accounts, 2},
     {:get_fees, 2},
     {:get_transfers, 2},
-    {:cancel_order, 3},
-    {:get_order, 3},
-    {:get_orders, 2},
     {:get_trade_history, 2},
     {:get_balances, 2},
     {:get_rate_limit_status, 2},
@@ -255,11 +252,15 @@ defmodule DpExchange.Webull do
   def replace_order(_credentials, _id, _request, _opts \\ []), do: Venue.not_supported()
 
   @impl true
-  def cancel_order(_credentials, _id, _opts), do: Venue.not_supported()
+  def cancel_order(credentials, client_order_id, opts \\ []),
+    do: Rest.cancel_order(credentials, client_order_id, with_limiter(opts))
+
   @impl true
-  def get_order(_credentials, _id, _opts), do: Venue.not_supported()
+  def get_order(credentials, client_order_id, opts \\ []),
+    do: Rest.get_order(credentials, client_order_id, with_limiter(opts))
+
   @impl true
-  def get_orders(_credentials, _opts), do: Venue.not_supported()
+  def get_orders(credentials, opts \\ []), do: Rest.get_orders(credentials, with_limiter(opts))
   @impl true
   def get_trade_history(_credentials, _opts), do: Venue.not_supported()